PP-120 Session state is keyed by charge ID
The session key 'ch_' + chargeId is now used to store things related to
that charge, allowing multiple charges to store data to the same session
cookie without accidentally stepping on each others toes.

Additionally, the handling of cookie code in the tests was repetitive
and a bit verbose, so this has been tidied up too.