PP-575 tidy up confirm endpoint
-hardcode url for the moment
we are following multiple different patterns for the endpoint urls, so it makes sense to standardise on one for the moment
everything else is hardcoded, so i went for that. we will be doing some middleware on the charge controller to
always get the charge, but we are not there yet. once we do this
it will be easier to switch to hateaos: