PP-441 make routes work for get and post
generating the routeName was not wokring as we had action that had the same
url for get and post (valid way of doing things)

to solve this problem we now in the paths say which type of route action it is
and use this in our code to figure out which named route it is

this has resulted in slightly more messyness with generateRoute, but not
currently 100% sure how to fix this. need some time to chew it over.

also added in basic state checking for the post views