PP-480 Start app as a cluster, with forever supervising master thread
Currently if there is an uncaught exception the whole app is liable to fall over.
This work adds two layers of resilience to prevent this occurring:
1. The node cluster api is used to run multiple threads, or workers.
These workers are supervised by the master thread, which restarts workers when they go bonk.
Number of workers currently configured to either read an environment variable NODE_WORKER_COUNT
or default to 1.

2. The master thread itself is invoked by forever https://github.com/foreverjs/forever
This will respawn the master thread if for some reason it exits. Forevert is run in the foreground,
as docker does not handle background processes very well.

This work also explicitly separates how app is run in development and production.
The development mode is the same as it was previously, with the forever + cluster
stuff only running when explicitly in production mode.