PP-1194 fix race condition on card id
the `var allowed` statement on line 9 of `card.js` declared a variable
in the scope of the module[1]

Because nodejs caches modules[2] when they are loaded, each module is
effectively a singleton. Therefore that `allowed` variable would be
shared between all invocations of the `checkCard` function, even when
requests are executing in parallel.

This fix removes the use of a module scoped variable, and instead passes
it in by binding it using a closure created in the module constructor
(`card.js:90`).

A further fix was required to `charge_validation_backend.js` which was
erroneously creating its own `Card` model instance, rather than using
the one passed to it in the `cardModel` variable.

with @heathd

[1] https://nodejs.org/api/modules.html#modules_the_module_wrapper
[2] https://nodejs.org/api/modules.html#modules_caching