PP-1335 Refactor out card id client from card model
In order to allow us to easily configure frontend http client
communication, we first need to extract out a common base client.
This commit is the first step, extracting out a base client and
card_id client from the card model. The keep-alive
configuration is added to the base client, as we expect all
clients to use keep-alive connections for internal calls.

This commit also sets maxSockets on the https globalAgent
to 100. This is believed to be necessary to ensure
connections are reused efficiently, and appears to be a good
figure based on preliminary testing.

with @PauloPortugal