PP-1335 Change base client implementation away from node-rest-client
Previous work was done to configure our http client library, node-rest-client,
to do keep-alive connections.

Under performance testing we did not see the desired behaviour given this configuration:
it appeared that frontend was still creating a new connection for every request.

This commit completely gets rid of node rest client and uses the underlying
node https library to make requests to downstream services, configured with
a custom agent that is configured to do keep-alives.