PP-1419 Fix validation without javascript
Previously if a user submitted an invalid payment details form without
javascript, they would be taken to /card_details, with invalid fields
highlighted. However on correcting and resubmitting they would see an error
as the chargeId was no longer contained in the url. This commit changes
the post path to match get path, so that user can resubmit form.