Move startup command to shell script
As part of https://github.com/alphagov/pay-infra/pull/699 we have
changed the way that we build docker containers which are deployed to
production. In particular we will be using `docker commit` to snapshot
an image from a container.

It turns out that when using `docker commit` it will use the last
command passed to `docker run` as the `Config/Cmd` setting (from docker
inspect output). This is normally provided by the `CMD` directive from
the `Dockerfile`[1], it can be over-ridden when running the container but
it seems like we don't do that. So, it's important that the `Config/Cmd`
setting is correct.

So, I'm standardising the startup command to invoke `docker-startup.sh`
so it's consistent with selfservice.

[1] http://goinbigdata.com/docker-run-vs-cmd-vs-entrypoint/