PP-1751 Updated product page to add support page
Updated `pay-product-page` to the latest version which includes the following changes:

Added a call to action on the product page and a support page that this call to action links to.
This support page sits within the product page domain and has information about how to get
in touch with Pay. The support link in the header also goes to this page.

https://github.com/alphagov/pay-product-page/pull/5