PP-1776 Compatibility fix to show relevant view on CAPTURE_APPROVED and CAPTURE_APPROVED_RETRY
As part of PP-1776, Connector will mark a charge as `CAPTURE_APPROVED` to record
the capture intend so that it can be picked up subsequently by a scheduled job. On the rare occasion
where the actual capture failed, Connector marks the charge as `CAPTURE_APPROVED_RETRY`
for later retrieval.
From a frontend point of view both statuses `CAPTURE_APPROVED` and `CAPTURE_APPROVED_RETRY`
should be treated as a successful capture from now on.

With this when the charge is `CAPTURE_APPROVED` or `CAPTURE_APPROVED_RETRY` or `CAPTURED`
the successful screen is shown.