PP-1804 Add validation for domain portion of email
Splits out the validation of an email address
string into its own function and adds a utility
module for validating parts of an address,
starting with a method for the domain part.

This begins the method with a full copy of the
regexes from the rfc822-validate library that we
use to match and validate complete addresses. The
next steps are to refactor this to only give us
the domain to operate on.