PP-2021 Add ability to use multiple session encryption keys
We need to rotate the `SESSION_ENCRYPTION_KEY` in frontend.
The `SESSION_ENCRYPTION_KEY` is used by the `client-sessions` library
to encrypt the session cookie that holds user state (most importantly the charge
id).
To rotate without disrupting payment s that already in flight we need to be able to use
multiple keys at once. The plan then would be to:

1. Add new key to environment `SESSION_ENCRYPTION_KEY_2`
2. With this change, app would then start storing user state on 2 cookies,
`frontend_state` and `frontend_state_2`
3. Leave to run in this dual mode for longer than `MAX_COOKIE_AGE`
(so that we know that all user state will exist in both cookies)
4. Remove (or rotate) `SESSION_ENCRYPTION_KEY`

 This commit reworks the existing cookie utility to allow this.

 The important methods on it are:

 `configureSessionCookie`: now configures `client-session` with all encryption keys in env
 `setSessionVariable`: now sets a variable on all session cookies for which an encryption key
 exists in env
 `getSessionVariable`: attempts to retrieve variable from either session cookie