PP-2276 Added integration tests and validation for too many non-consecutive numbers *What* - Added integration tests for 'too_many_digits' validation - Altered 'too_many_digits' validation to check for non-consecutive numbers so we can catch potential space or dash separated PANs in non-PAN fields - Updated unit tests accordingly
*WHY*
To mitigate risk of re-occurrence of PANs ending up being stored in DB/logs. When this last happened the query below had to be used to identify polluted records:

```
select  created_date, id, address_line1, regexp_replace(address_line1, '[0-9]', '*', 'g' ) from charges where address_line1 ~ '.*[0-9].*[0-9].*[0-9].*[0-9].*[0-9].*[0-9].*[0-9].*[0-9].*[0-9].*[0-9].*[0-9].*';
```

This is equivalent to counting the number of digits in a field. We're using the below to check for this in future:
```
function hasTooManyDigits(input) {
    if (!input || typeof input !== 'string') return false;
    const matchedDigits = input.match(/(\d)/g);
    return (matchedDigits !== null) && (matchedDigits.length >= 10) ;
  }
```