PP-2377 - Integrated accessible autocomplete
I’ve added [accessible autocomplete](https://github.com/alphagov/accessible-autocomplete) to progressivly enhance the select box we use on the payment form.

It’s pulled in with Browserify and I’ve added grunt-contrib-cssmin which concats the aut
ocomplete styles onto our application stylesheet AND now we're minifying the CSS too whi
ch is saving 5kb 🎷

Also updated validation because utofill prefills the hidden selected country option, emptying the field doesnt result in error about not choosing a country because you still have that country selected as no other one was chosen instead. So on validation I have made it refill out the autocomplete with the country that is selected to make it clear why it isnt erroring.