Remove package-lock.json
WHAT:

Currently we are using Node 6.11.1 and package-lock.json is ignored.
When we upgrade to Node 8 it will be used by default (if there is `package-lock.json` and `package.json`, then `package-lock.json` is used in favour of `package.json`).
When we upgrade to Node 8 we should keep in sync both `package-lock.json` and `package.json`.

WHY:

We don't want to use currently outdated `package-lock.json` and we want to add it again once we integrate Node 8.