PP-3022 - Wrapping input/selects in their label element causes accessibility issues
Whilst it’s valid to wrap `input` or `select` within a `label` it causes issues for
screen reader uses on PC. The when focusing the label element it will read out
everything within the label and so will read the entire select box, in our case
this is a list of 200+ countries which means users probably wont listen to the
end and find that its a combo box that can be accessed.

So I've moved the inputs/selects outside of the labels across the whole page.