PP-2992 Nunjucks migration - npm packages
In order to get ready for Nunjucks, we need to remove the old template engine, Hogan. So we have removed the packages and the shims to initalise the engine.

Secondly, we have removed the mustache template as we will be replacing with a Nunjucks version.

And finally we added Nunjucks library