PP-3302 - using our countries list not Registers
We’re using our list because we know it works with our
payment providers, there were 31 differences between the two
they were mostly extra territories or countries that no longer
exist. This means I also had to update the graph file which maps
synonyms because it will make it appear like we accept countries
that are not actually in our list.