PP-3453 Handle EPDQ response for 3DS
- When receiving a response from the payment gateway, it will
send the user back inside the iframe. To get out from the iframe
we use a form with a target `_top` and at the same time we pass
any data we get from the payment provider as form fields to the
charge_controller. Because some of this data is specific to each
payment gateway, we will set it only if available. After setting this
data as hidden fields we then submit the form and send this data
to connector
- Added both POST and GET to the same path for EPDQ as they
use GET as a fall back when JS is deactivated. This callback url to us
will contain the EPDQ auth status as query param
- The status received from EPDQ is not reliable as this is
resubmitted as a hidden field, however we verify this status in connector
against EPDQ before updating the charge

- Updated auth_3ds template to hide/display data for EPDQ or
WorldPay
- Updated routes for handling EPDQ 3DS
- Added tests

with @kakumara