BAU - The email typo thing needs more work
It’s overly sensitive and is flagging more false positives than actual
errors, need to roll back and work out how to make it less trigger
happy.

On the plus side, we were noticing it dropping the persons  email from
the form on error, this was not because of mailcheck but due do a
erroeous deletion during a rebase of the Node 8 upgrade. So that’s one
'bug' fixed :D