PP-3133: Move routes init to after public
In the previous commit we added a catchall route, which redirects to the
GOV.UK 404 page. However public/static assets were initialised after
routes were initialised. This meant that no static assets were served
but instead were served as 404s.

Moving the initialisation of static assets to before the initialisation
of routes means that the 404 catchall is the last route registered and
thus will only match things that should 404.

solo @tlwr