PP-4090 - Switch for ZAP to disable JS Validation
ZAP tests need to be able to submit a malformed post so that it can run
a bunch of pen tests on the post. But to do that, it needs to get
around our client-side validation JS. Before they could use jQuery to
kill all the eventListeners but we can’t do that without jQuery.

So I’ve put a breaker in the function that will supress the validation
if an element with the id disable-for-zap is present.

Then when the page is loaded in the ZAP test it will insert the element
before it does the submit and will be able to do the pen test.