bau: log memory stats
As part of the performance testing epic
(https://payments-platform.atlassian.net/browse/PP-4004) we have made some
changes to improve performance Some of these are:

- caching token requests in publicapi
- upgrading frontend's ec2 instances from m3.medium -> c4.large

It would be good to log the memory statistics for each server process in node
should we need to further improve performance (for example, by changing the
heap size, which would then affect when the garbage collection kicks in) for
frontend in the future. We currently have no easy way to know this.

@oswaldquek