PP-4365 - Get started with connecting Apple Pay to connector
Solo @simad (commit message posthumously written by @jonheslop)

Just to begin with introducing hardcoded stub on apple pay encrypted
payload for test integration purposes.

This payload has two parts, the encrypted bit we don’t mess with and
the payment info (name, email, last 4 digits, card brand etc) that
needs to be normalised to bring inline with how connector
expects it to be.

This commit introduces new endpoint that the browser submits the
payload to and then the normalisation to get it ready for connector.