PP-4896: Domain association endpoint for Apple Pay
This commit adds a route, controller and relevant tests for a new
`/.well-known/apple-developer-merchantid-domain-association.txt`
endpoint. This endpoint is used by Apple Pay to verify that our
Merchant ID controls a particular Domain Name.

Apple provide us with a text file to serve from the endpoint. This
text is environment-specific and appears to be fetched periodically,
so we need  to pass it in as an environment variable rather than
hardcoding the value.