PP-4914 - Apple need multiline certs
Due to limitations with our current infra set up all secrets need to be
single line.

However Apple requires multiline serts and keys to be able to validate
us as merchant.

Whilst we’re working on making multiline secrets poosible we’re
manipulating the secrets in the controller to get around the problem.