PP-4917 - Validate users email to check its legit
Apple let users put anything in the email field, this makes sure its
legit.

I cant try this locally so I have no idea if it will work (and the docs
are rubbish) so would like to merge and try on test to see if it works.

Hence why I have left a log line because i imagine this wont work and
this will help me debug

Also spotted an abort that was causing error in certain cases
as it shouldnt have been there