Add common way to setup logger
At the moment, different files and components are loading `winston`
directly, which means the setup would have to be duplicated in many
places or potentially diverge.

This isn't cool. Adding a function that will be loaded from everywhere
instead, gives us the lever of abstraction we need without refactoring
the app to use context.