PP-5191 Allow auth success for auth handler
With stripe 3ds authorisations, the actual authorisation is triggered by
a notification, so has nothing to do with what the browser is up to. For
this reason, if there is some network slowness you can end up with the
user being returned after the payment has been authorised. Currently
this ends up with the user having the 'payment in progress' screen
displayed, because the state enforcer did not like that combination of route
and status. This commit allows this combination.

AS part of this I've deleted one of the assertions in the state service
test, as it no longer made sense. The particular functionality it was testing
is never used.