PP-5279 - fix Apple Pay when Google Pay is also availabled
If both are available we had overrode the Apple Button with a Google one.

Now we make a separate form for each and use Javascript/CSS to determine which one to present to user.

Had to also rejig the JS to understand there are two forms on the page and it needs to set event listeners on both