PP-5366 - pattern is problematic
Whilst it makes the number pad show on older iOS devices also then show
native error pop ups if the users uses non whitelisted characters we
want to allow people to enter their card number with or without spaces.

Editing the regex to allow spaces stops it from showing the number pad
so it’s redundant to include it.