RED-116: Use tini in docker to run node
By having bash as PID 1, we have a problem. Signals sent to the container (e.g.
SIGTERM to shut it down) will not reach npm - bash will not pass them on. This
means that shutting the container down is done uncleanly. Once the SIGTERM is
ignored, the container orchestrator will then send a SIGKILL and forcibly exit
the process.

We could make npm PID 1 instead but this creates further problems - If npm
creates any child processes and those processes (or any of their children) die,
npm will be the parent of them and will not know to waitpid() on them, so they
will remain zombies forever.

tini solves these problems - it will correctly reparent orphaned child
processes and pass signals through to npm.