PP-5546 Add pact test for find charge
Add pact test for GET charge connector endpoint.
Include the new 'integration_version_3ds' property of the gateway
account.

Modify pact_base so that it will add matchers for objects in nested
arrays so that we can assert on the structure of 'card_types' entries.

Remove fields from charges fixture which are not used in frontend.