PP-5548 - Move field validation into browsered scripts
Because we later want to hook onto form validations submit event hijack, we want to be able to require another JS file.

It’s better to do this with browserify so we needed to move `form-validation.js` to the browsered folder so it’s transformed by Grunt.

In moving it we had to change when and how the validation is initiated. This has moved from `scripts.njk` to `browsered.js`

This changed when the script is initiated but the validation needs the country picker to be ready so we had to flip the order of loading there too, this was done inside `charge.njk`