PP-5550: Minor simplification of charge_controller
Writing a unit test for PP-5550 is complicated. Possible - but complicated.
This is because the charge_controller's "create" method is unwieldy, too big
and overly complicated. Shockingly there are no unit tests for the method.
Unsurprising given the complexity of the code.

In order to write a unit test for PP-5550 I've decided to make small
incremental simplifications and eventually move the validation code into its
own class. But for now, this method gets rid of the 'let userEmail' as its
result is never used.

I have also taken a decision - in conjunction with @mrlumbu and @sfount - to
get rid of some AWSXray code. It makes code look more complicated; and in
reality we use request ids and charge ids to trace a request through the
system.