PP-5758 Change state enforcer log line to log at INFO level
We log when the state enforcer decides that the charge is not in the
correct state to access a route, and as a result render an appropriate
page. This is business as usual, and part of the design of connector,
so we shouldn't log this at ERROR level, instead log at INFO level.

Co-authored-by: Dan Worth <dan.worth@digital.cabinet-office.gov.uk>