BAU - Remove trailing dashes in log lines
These dashes were there previously because before we logged in JSON the
logger would concatenate the JSON object it was passed as a second
parameter into the line.

Now, however, the keys in this JSON object are merged into the top-level
JSON object for the logline, so there is a hanging dash on the end of
the message which makes it appear that the message is incomplete. Remove
these.