PP-5887: add zap test hash to csp depending on configuration
Why?
When enabling csp in enforce mode for end2end tests, the zap tests are failing
because they try to inject the script into the frontend page (which csp is preventing
from doing).
It seems we're not allowed to simply disable javascript while running selenium tests,
so in order to workaround the issue with injectint the script we're going to allow it,
but only when specified in environment variable (it will be set only for end to end tests)

Changes:
* add hash to scriptSrc derective if it's set up in environment variable
* add test to cover the new scenario