PP-5762 Add common keys to logs
Use continuation-local-storage to store values for keys that should be
included in log lines when they are available in the middleware. Then
in the logger, retrieve these values and add them to the logs.

We are using the continuation-local-storage package rather than
cls-hooked as there have been some issues with async_hooks since our
recent upgrade to node 12.

Use the namespace already used for storing the AWS xray segment to store
the logging key values, and move the middleware that creates this
namespace and binds it to the callback chain into its own file.

Add the following keys for now:
x_request_id (the request correlation id)
payment_external_id