PP-5762 Add common keys to logs
Add common logging fields to an object on the request and retrieve them
when we do logging to include them in the structured logs.

Added the `payment_external_id`, `gateway_account_id`,
`gateway_account_type` and `provider` fields.

Included one example of getting these fields from the request and
passing them to the logger.

Originally considered using either the `continuation-local-storage` or
`cls-hooked` packages to achieve storing key/values for a request and
retrieving them when we did the logging. This would have been nicer,
however these seemed to be too unreliable. `continuation-local-storage`
does not work with async/await and `cls-hooked` uses async hooks which
are considered experimental and were broken in a recent version of node.