PP-5762 Log fields added to request object
In all places where we do logging, retrieve the logging fields that
we've added to the request and pass them to the logger.

As part of doing this, have removed the `logging.systemError` method
as the purpose of this was to log certain values in the error which are
now added to the logging fields stored on the request.