PP-5959 CSP: country autocomplete needs script-src 'unsafe-inline'
The country and territory autocomplete seems to require
'unsafe-inline' to be listed as one of the sources for the
style-src Content Security Policy directive.

We believe this is because the accessible autocomplete component
upon which it is built tries to set the style attribute on an
element when it detects it is running in an iOS browser.

We have filed an upstream issue:
https://github.com/alphagov/accessible-autocomplete/issues/398