PP-6079 Don't collect billing address for moto payments
Currently, we don't show fields to collect the billing address on the
card details page if billing address collection is disabled for the
service.

Now, we also won't show these fields if a payment is a MOTO payment even
if billing address collection is enabled for the service. Determine
whether we should collect the address in the resolve_service middleware
and add a field to res.locals so that we can access it in all templates
and controllers that need this value.

Add Cypress tests to test this and remove tests from charge_ui_tests
which are now better covered by Cypress tests.