PP-6147: Middleware to decrypt fields
When enabled (i.e. when `DECRYPT_AND_OMIT_CARD_DATA` is `true`), this
middleware will decrypt `req.body.cardNo` and the other card data
fields.

When `DECRYPT_AND_OMIT_CARD_DATA` is set, all other configuration
(private key, key name and namespace) is mandatory, and AWS CloudFront
MUST be configured to encrypt the data - if unencrypted card data is
sent to frontend while this feature is enabled it will result in an
error.