PP-6147: Remove async from decryption middleware
We use a dependency called continuation-local-storage, which attempts
to store data, count how many functions have been started and finished
since the data was written, and keep the data around until the count
reaches zero.

This gets confused if a piece of middleware calls `next()` AND resolves
a promise (as this was doing, thanks to the async / await syntax).

The result is that when we later attempted to read from
continuation-local-storage the data we want wasn't there.

We're only using continuation-local-storage for the AWS xray segment,
which we already have a story to remove: [PP-5805](https://payments-platform.atlassian.net/browse/PP-5805)

Note: I've kept the async / await syntax in the tests because it makes
them much easier to read / maintain compared to putting expectations in
`.then()` callbacks.