PP-6147: form-url-encode card type ajax call
AWS CloudFront's Field Level Encryption feature only works for
application/x-www-form-urlencoded content. Because this XHR sends
cardholder data, we want CloudFront to encrypt it, so we have to use
this content type. Express is already configured to use bodyParser to
parse form-encoded bodies, so there's no need for any magic on the
server side, other than using the decryptCardData middleware to decrypt
the cardNo.

Also fixes a dodgy call to `logger.error()` (it expects an object as a
second argument, not a string, causing it to log
`{"0":"f","1":"i","2":"e","3":"l","4":"d","5":"N","6":"a","7":"m","8":"e"}`,
which is hard to read to say the least.)

There are no unit tests covering this code - it seemed
like overkill to add one.

**Note:** this is not feature flagged, as it should work fine in
production without a flag - we should do a bit of testing to confirm
that though.