PP-6147: Handle middleware errors in tests
Previously the promise we're using in the tests (so we can use async /
await syntax) would always resolve, even if the middleware called `next`
with an error parameter.

Forcing the code to error confirms that the tests weren't doing what
they were supposed to (they all pass, even when the code is broken).

Inspecting the `err` parameter and rejecting the promise if it's truthy
means the tests behave properly.