PP-6259 Make frontend send user to do 3DS again if necessary
When frontend calls connector to authorise a 3D Secure result with
the payment gateway, redirect the user to the 3DS Secure page
again if connector’s response indicates that this is required.

Connector indicates this by sending a response body of {"status":
"AUTHORISATION 3DS REQUIRED"} when /v1/frontend/charges/{chargeId}
/3ds is called (this isn’t actually implemented in connector yet
but will be soon).

with @SandorArpa