PP-5805: Remove aws xray from tests
As part of this commit the following needed to be added to charge_ft_tests and
charge_validation_ft_tests

```
const app = proxyquire('../../server.js',
  {
    'memory-cache': {
      get: function () {
        return false
      },
      '@global': true
    }
  }).getApp()
```
This mocks out the memory-cache in the resolve_service
(https://github.com/alphagov/pay-frontend/blob/master/app/middleware/resolve_service.js#L4).

It turns out:
- most of the tests across the project uses the same gatewayAccountId
- there is some test somewhere (I don't know where) that associates a service
  to this gateway account which sets collectBillingAddress to false

This caches the service and  has a knock on effect on the rest of the tests
where the billing address needs to be collected and causes the following code
in the charge_controller to be executed
https://github.com/alphagov/pay-frontend/blob/master/app/controllers/charge_controller.js#L187-L189:
```
    if (res.locals.collectBillingAddress === false) {
      delete payload.address
    }
```

This causes the tests to fail because the expectation at
https://github.com/alphagov/pay-frontend/blob/master/test/integration/charge_ft_tests.js#L104-L116
fails.

Some solutions:
- mock out the memory-cache globally (the chosen solution)
- make every gateway account id random so as to bypass the cache in the
  resolve_service (too much effort)
- set the SERVICE_CACHE_MAX_AGE which is used in resolve_service to 0 (hard to
  do)