PP-6676 Set node version for concourse build packs
The highest version of 12.X node that pass build pack supports is 12.16.3 so
set the node engine version to use this or higher. This setting is only
used by PaaS buildpack at present since Jenkins Ci will reference .nvmrc
and the app is deployed to current infra in a container with node
version set via the docker image base.