PP-6706 Make example card expiry date always be in the future
Make the example card expiry date on the enter card details page
always use a year two years ahead of the current year in the UK.

When it’s 2020 in the UK, the example card expiry date will be
10/22. When it’s 2021 in the UK, the example card expiry date will
be 10/23. When it’s 2022 in the UK, the example card expiry date
will be 10/24. And so on. (The month always remains 10.)

This means that the example card expiry date will be between 1 year
10 months and 2 years 10 months in the future, which is appropriate
given that most cards expire between 3 and 5 years after they are
issued.

The example year will change just once each year at the midnight
between New Year’s Eve and New Year’s Day in the UK.